refactor(composite): add maybeAnimation
