+++ src/geo/geoView.ts
@@ -1,7 +1,7 @@
 import { Linear } from '@antv/scale';
 import { upperFirst } from '@antv/util';
 import { geoPath, geoGraticule10 } from 'd3-geo';
-import { maybeTooltip } from '../utils/tooltip';
+import { maybeTooltip } from '../utils/mark';
 import { CompositionComponent as CC } from '../runtime';
 import { GeoViewComposition } from '../spec';
