+++ src/mark/treemap.ts
@@ -12,7 +12,7 @@ import { subObject } from '../utils/helper';
 import { CompositionComponent as CC } from '../runtime';
 import { TreemapMark } from '../spec';
 import { getBBoxSize } from '../utils/size';
-import { maybeTooltip } from '../utils/tooltip';
+import { maybeTooltip } from '../utils/mark';
 import { generateHierarchyRoot, field } from './utils';
 
 export type TreemapOptions = Omit<TreemapMark, 'type'>;
